Fixed to handle LeaveNotify.
GIMP seems to define no cursor (XUndefineCursor) on the west scale. So
the frame's cursor is used (if the pointer moves from the west edge to
GIMP's west scale, the cursor is kept as the left bar "|<-").

When the pointer enters into the client area, the parent frame accepts
LeaveNotify event. It is time to XUndefineCursor for the parent frame.