Fixed searching a frame for ConfigureRequest.
The previous code expected that a parent member of a message is always
one of frame windows.

But I guess that an X application sends messages before its window is
reparented. So an window manager must handle messages whose parent
member is not a frame window, but a root window, even after an window of
messages was reparented.

So I changed process_configure_request() to search a frame by child
window handle.